Move code related to FullSync to full_sync.rs
